influx: fix failing test case:
This commit illustrates shows, IMO, how succinctly you can change the
whole state mutation story using influx.  We went from having the
Operator action be a dumb assignment (with crufty display management)
to have it actually triggering a calculation. additionally, the
plus/minus operator changed in a significant way.

The code is roughly the same complexity of the previous version, but
also improves the usability of the calculator immensely (and fixes a
test case).